+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Datapack support for custom mob drops
 ### Changed
 - Override of mob specific configuration (policy.mob) has been made more user friendly
+- Fixed Anvil JEI recipe not showing controller recipe [#429]
 ## Removed
 - Stuff that has been removed